# perf: consume less CPU cycles, by understanding tokio